move file to shared
